# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TopItem {
  name: string;
  count: number;
  revenue: number;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  averageOrderValue: number;
  revenueChange: number;
  ordersChange: number;
  customersChange: number;
  topItems: TopItem[];
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  items: string;
  total: number;
  status: string;
  orderType: string;
  createdAt: string;
}

export interface OrderUpdate {
  status?: string;
}

export interface MonthlyReport {
  month: string;
  revenue: number;
  orders: number;
  averageOrderValue: number;
  newCustomers: number;
}

export interface CategoryStat {
  category: string;
  revenue: number;
  orders: number;
  percentage: number;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  totalOrders: number;
  totalSpent: number;
  lastOrderDate: string;
  status: string;
  joinedAt: string;
}

export type ListOrdersParams = {
status?: string;
limit?: number;
offset?: number;
};

export type ListCustomersParams = {
search?: string;
limit?: number;
offset?: number;
};