# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: Admin dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard overview stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/orders:
    get:
      operationId: listOrders
      tags: [admin]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/monthly:
    get:
      operationId: getMonthlyReport
      tags: [admin]
      summary: Get monthly sales report
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyReport"

  /admin/orders/by-category:
    get:
      operationId: getCategoryStats
      tags: [admin]
      summary: Get sales breakdown by category
      responses:
        "200":
          description: Category stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryStat"

  /admin/orders/recent:
    get:
      operationId: getRecentOrders
      tags: [admin]
      summary: Get recent orders for dashboard feed
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/{id}:
    patch:
      operationId: updateOrder
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  /admin/customers:
    get:
      operationId: listCustomers
      tags: [admin]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"

  /admin/customers/{id}:
    get:
      operationId: getCustomer
      tags: [admin]
      summary: Get single customer with order history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AdminStats:
      type: object
      required:
        - totalRevenue
        - totalOrders
        - totalCustomers
        - averageOrderValue
        - revenueChange
        - ordersChange
        - customersChange
        - topItems
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalCustomers:
          type: integer
        averageOrderValue:
          type: number
        revenueChange:
          type: number
        ordersChange:
          type: number
        customersChange:
          type: number
        topItems:
          type: array
          items:
            $ref: "#/components/schemas/TopItem"

    TopItem:
      type: object
      required: [name, count, revenue]
      properties:
        name:
          type: string
        count:
          type: integer
        revenue:
          type: number

    Order:
      type: object
      required:
        - id
        - customerName
        - customerPhone
        - items
        - total
        - status
        - orderType
        - createdAt
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        items:
          type: string
        total:
          type: number
        status:
          type: string
        orderType:
          type: string
        createdAt:
          type: string

    OrderUpdate:
      type: object
      properties:
        status:
          type: string

    MonthlyReport:
      type: object
      required:
        - month
        - revenue
        - orders
        - averageOrderValue
        - newCustomers
      properties:
        month:
          type: string
        revenue:
          type: number
        orders:
          type: integer
        averageOrderValue:
          type: number
        newCustomers:
          type: integer

    CategoryStat:
      type: object
      required:
        - category
        - revenue
        - orders
        - percentage
      properties:
        category:
          type: string
        revenue:
          type: number
        orders:
          type: integer
        percentage:
          type: number

    Customer:
      type: object
      required:
        - id
        - name
        - phone
        - totalOrders
        - totalSpent
        - lastOrderDate
        - status
        - joinedAt
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lastOrderDate:
          type: string
        status:
          type: string
        joinedAt:
          type: string