# lib/api-zod/src/generated/types/monthlyReport.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyReport {
  month: string;
  revenue: number;
  orders: number;
  averageOrderValue: number;
  newCustomers: number;
}