# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  items: string;
  total: number;
  status: string;
  orderType: string;
  createdAt: string;
}